+++ src/main/kotlin/minesweeper/domain/Cell.kt
@@ -1,0 +1,3 @@
+package minesweeper.domain
+
+class Cell(val bomb: Boolean = false)
