+++ src/main/kotlin/minesweeper/domain/Cells.kt
@@ -1,0 +1,3 @@
+package minesweeper.domain
+
+class Cells(val cells: List<Cell>, val width: Int) : List<Cell> by cells
