+++ src/main/kotlin/minesweeper/main.kt
@@ -1,0 +1,4 @@
+package minesweeper
+
+fun main() {
+}
