+++ src/test/kotlin/minesweeper/view/BoardViewTest.kt
@@ -1,12 +1,10 @@
 package minesweeper.view
 
-import minesweeper.domain.MotherCellTest.Cell
-import minesweeper.domain.MotherCellTest.Cells
+import minesweeper.domain.Cell
+import minesweeper.domain.Cells
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import java.io.PrintWriter
 import java.io.StringWriter
-import java.io.Writer
 
 class BoardViewTest {
     @Test
@@ -30,30 +28,5 @@ class BoardViewTest {
             ◻️💣
             """.trimIndent()
         )
-    }
-
-    class BoardView(private val cells: Cells, private val out: PrintWriter) {
-        constructor(cells: Cells, writer: Writer) : this(cells, PrintWriter(writer, true))
-
-        fun show() {
-            val height = cells.size / cells.width
-            val result = (1..height).joinToString(System.lineSeparator()) {
-                val startIndex = (it - 1) * cells.width
-                val endIndex = startIndex + cells.width
-                cells.subList(startIndex, endIndex).map(::CellView).joinToString("")
-            }
-            out.print(result)
-        }
-    }
-
-    class CellView(private val cell: Cell) {
-        override fun toString(): String {
-            return if (cell.bomb) BOMB else BLANK
-        }
-
-        companion object {
-            const val BLANK = "◻️"
-            const val BOMB = "💣"
-        }
     }
 }
